Bugfix for periodic boundary conditions
It looks like the old version tried to avoid unnecessary
evaluations of the flow by only applying the flow to the
first `num_real_points` points. The actual redundant points
however, may be distributed differently across the indices.
Also, in adaptiveTOFutureGrid it is assumed that `flow_map` is
populated with the images of every point.